New package for working copy related classes
