+++ scm-core/src/main/java/sonia/scm/repository/api/RepositoryServiceFactory.java
@@ -53,7 +53,7 @@ import sonia.scm.repository.RepositoryManager;
 import sonia.scm.repository.RepositoryPermissions;
 import sonia.scm.repository.spi.RepositoryServiceProvider;
 import sonia.scm.repository.spi.RepositoryServiceResolver;
-import sonia.scm.repository.util.WorkdirProvider;
+import sonia.scm.repository.work.WorkdirProvider;
 import sonia.scm.security.ScmSecurityException;
 
 import java.util.Set;
