+++ scm-core/src/main/java/sonia/scm/repository/work/WorkdirProvider.java
@@ -22,45 +22,32 @@
  * SOFTWARE.
  */
 
-package sonia.scm.repository.util;
-
-import sonia.scm.util.IOUtil;
+package sonia.scm.repository.work;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.Files;
 
-public final class WorkingCopy<R, W> implements AutoCloseable {
+public class WorkdirProvider {
 
-  private final File directory;
-  private final W workingRepository;
-  private final R centralRepository;
-  private final Runnable close;
+  private final File rootDirectory;
 
-  public WorkingCopy(W workingRepository, R centralRepository, Runnable close, File directory) {
-    this.directory = directory;
-    this.workingRepository = workingRepository;
-    this.centralRepository = centralRepository;
-    this.close = close;
+  public WorkdirProvider() {
+    this(new File(System.getProperty("scm.workdir" , System.getProperty("java.io.tmpdir")), "scm-work"));
   }
 
-  public W getWorkingRepository() {
-    return workingRepository;
+  public WorkdirProvider(File rootDirectory) {
+    this.rootDirectory = rootDirectory;
+    if (!rootDirectory.exists() && !rootDirectory.mkdirs()) {
+      throw new IllegalStateException("could not create pool directory " + rootDirectory);
+    }
   }
 
-  public R getCentralRepository() {
-    return centralRepository;
-  }
-
-  public File getDirectory() {
-    return directory;
-  }
-
-  @Override
-  public void close() {
-    close.run();
-  }
-
-  void delete() throws IOException {
-    IOUtil.delete(directory);
+  public File createNewWorkdir() {
+    try {
+      return Files.createTempDirectory(rootDirectory.toPath(),"workdir").toFile();
+    } catch (IOException e) {
+      throw new RuntimeException("could not create temporary workdir", e);
+    }
   }
 }
