--- scm-plugins/scm-git-plugin/src/test/java/sonia/scm/repository/CloseableWrapperTest.java
@@ -1,55 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020-present Cloudogu GmbH and Contributors
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-    
-package sonia.scm.repository;
-
-import org.junit.Test;
-import sonia.scm.repository.util.CloseableWrapper;
-
-import java.util.function.Consumer;
-
-import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.verify;
-
-public class CloseableWrapperTest {
-
-  @Test
-  public void shouldExecuteGivenMethodAtClose() {
-    Consumer<AutoCloseable> wrapped = new Consumer<AutoCloseable>() {
-      // no this cannot be replaced with a lambda because otherwise we could not use Mockito#spy
-      @Override
-      public void accept(AutoCloseable s) {
-      }
-    };
-
-    Consumer<AutoCloseable> closer = spy(wrapped);
-
-    AutoCloseable autoCloseable = () -> {};
-    try (CloseableWrapper<AutoCloseable> wrapper = new CloseableWrapper<>(autoCloseable, closer)) {
-      // nothing to do here
-    }
-
-    verify(closer).accept(autoCloseable);
-  }
-}
