+++ scm-plugins/scm-hg-plugin/src/main/java/sonia/scm/repository/spi/HgWorkingCopyFactory.java
@@ -26,7 +26,7 @@ package sonia.scm.repository.spi;
 
 import com.aragost.javahg.Repository;
 import com.aragost.javahg.commands.PullCommand;
-import sonia.scm.repository.util.WorkingCopyFactory;
+import sonia.scm.repository.work.WorkingCopyFactory;
 
 public interface HgWorkingCopyFactory extends WorkingCopyFactory<Repository, Repository, HgCommandContext> {
   void configure(PullCommand pullCommand);
